fixed if statements jump not occurring

+++ src/debug.rs
@@ -52,9 +52,9 @@ pub fn disassemble_instruction(chunk: &Chunk, offset: u8) -> u8 {
 }
 
 fn jump_instruction(name: &str, chunk: &Chunk, sign: i8, offset: u8) -> u8{
-    let jump: usize = usize::from(chunk.code[offset as usize + 1] << 8 + chunk.code[offset as usize + 2]);
+    let jump: usize = (usize::from(chunk.code[offset as usize + 1]) << 8) + usize::from(chunk.code[offset as usize + 2]);
 
-    print!("{} {} -> {}", name, offset, offset as usize + 3 + sign as usize * jump);
+    print!("{} {} -> {}", name, offset, (offset as usize) + 3 + (sign as usize) * jump);
     offset + 3
 }
